Clearing up more unused code in NEURON and added implementations for reverse_permute
